refactor(travel): dedupe confirmed-event projection + city-name split
Two small cleanups flagged by prior review loops:

1. **`toExportableConfirmedEvent`** — the 9-line projection from
   `SerializedConfirmed` to `ExportableConfirmedEvent` was duplicated
   byte-for-byte across TravelResultsServer and SavedTripPage. New
   helper lives at `src/lib/travel/export.ts` alongside the
   canonical `ExportableConfirmedEvent` type. TripSummary re-exports
   the type for its existing public surface.

   Param typed as `Pick<SerializedConfirmed, …>` (not inline) so
   future additions to `ExportableConfirmedEvent` surface a compile
   error at the projection and force call sites to opt in.

2. **`extractCityName`** — `label.split(",")[0]?.trim() ?? label`
   was repeated at 4 sites: `ItineraryHero`, `MultiStopHeader`, the
   SavedTripCard delete dialog, and `DestinationBoardingStamp` in
   TravelResults. New helper in `@/lib/travel/format` colocated with
   `cityToIata` (same split convention).

**Pre-PR codex adversarial review caught a real RSC boundary hazard**:
initial draft put `toExportableConfirmedEvent` in `TripSummary.tsx`
(`"use client"` module), which server `page.tsx` then imported — a
classic RSC boundary violation that `tsc` accepts but `next build`
would reject. Fixed by lifting the helper into the plain
`@/lib/travel/export.ts` module; `npx next build` now passes clean.
